Added JDABuilder presence setters and improved handling of initial presence information for client-accounts
